Remove string format warnings by cleaning up usage of NSInteger and int. Declaring NSInteger where possible to keep casting on a low level.

+++ bibapp/BADetailScrollViewDelegate.h
@@ -11,6 +11,6 @@
 @protocol BADetailScrollViewDelegate <NSObject>
 
 - (void)continueSearch;
-- (void)updatePosition:(int)updatePosition;
+- (void)updatePosition:(NSInteger)updatePosition;
 
 @end
